move DOM traversal into electron package for now
