--- packages/dom/README.md
@@ -1,11 +1,0 @@
-# dom
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build dom` to build the library.
-
-## Running unit tests
-
-Run `nx test dom` to execute the unit tests via [Jest](https://jestjs.io).
