--- packages/dom/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@bitdrift/dom",
-  "version": "0.0.1",
-  "private": true,
-  "type": "commonjs",
-  "main": "./src/index.js",
-  "types": "./src/index.d.ts",
-  "dependencies": {
-    "tslib": "^2.3.0"
-  }
-}
