--- packages/dom/project.json
@@ -1,19 +1,0 @@
-{
-  "name": "@bitdrift/dom",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/dom/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/packages/dom",
-        "main": "packages/dom/src/index.ts",
-        "tsConfig": "packages/dom/tsconfig.lib.json",
-        "assets": ["packages/dom/*.md"]
-      }
-    }
-  }
-}
