+++ packages/electron/src/lib/renderer/index.ts
@@ -1,0 +1,2 @@
+export * from './addSessionReplayCapture';
+export * from './autoExposeInMainWorld';
