+++ packages/electron/src/lib/renderer/replay/captureScreen.ts
