+++ packages/electron/src/lib/renderer/replay/colorParse.spec.ts
