+++ packages/electron/src/lib/renderer/replay/colorParse.ts
