+++ packages/electron/src/lib/renderer/replay/constructArrayBuffer.ts
