+++ packages/electron/src/lib/renderer/replay/index.ts
@@ -5,5 +5,5 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
-export * from './lib/captureScreen';
-export * from './lib/constructArrayBuffer';
+export * from './captureScreen';
+export * from './constructArrayBuffer';
