+++ packages/electron/src/lib/renderer/replay/utils.ts
